functions/queue-processor/requirements.txt
azure-servicebus
azure-cosmos
